+++ src/task.js
@@ -1,7 +1,8 @@
-import {createElement} from "./create-element";
+import {Component} from "./component";
 
-export class Task {
+export class Task extends Component {
   constructor(data) {
+    super();
     this._title = data.titles;
     this._dueDate = data.dueDate;
     this._tags = data.tags;
@@ -10,8 +11,8 @@ export class Task {
     this._repeatingDays = data.repeatingDays;
     this._isDone = data.isDone;
     this._isFavorite = data.isFavorite;
+    this._onEditButtonClick = this._onEditButtonClick.bind(this);
 
-    this._element = null;
     this._onEdit = null;
   }
 
@@ -25,9 +26,6 @@ export class Task {
     }
   }
 
-  get element() {
-    return this._element;
-  }
   set onEdit(fn) {
     this._onEdit = fn;
   }
@@ -246,21 +244,11 @@ export class Task {
 
   bind() {
     this._element.querySelector(`.card__btn--edit`)
-        .addEventListener(`click`, this._onEditButtonClick.bind(this));
+        .addEventListener(`click`, this._onEditButtonClick);
   }
 
   unbind() {
     this._element.querySelector(`.card__btn--edit`)
-        .removeEventListener(`click`, this._onEditButtonClick.bind(this));
-  }
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
+        .removeEventListener(`click`, this._onEditButtonClick);
   }
 }
